+++ src/services/v2-service.js
@@ -37,7 +37,20 @@ export function submitOtp(id, code, trustDevice) {
 		body: JSON.stringify({
 			id,
 			code,
-			unsecure_device: !trustDevice
+			unsecure_device: trustDevice
 		})
 	});
+}
+
+export function generateOtp(labelPrefix) {
+	return fetch(`/zx/auth/v2/otp/generate?labelPrefix=${encodeURIComponent(labelPrefix)}`, {
+		method: 'POST',
+		headers: {
+			'Content-Type': 'application/json'
+		},
+		body: JSON.stringify({
+			labelPrefix,
+			humanReadable: false
+		})
+	}).then((res) => res.json());
 }